Refactored all typing files into /types folder
