+++ src/modules/api/user.ts
@@ -17,8 +17,8 @@
 */
 
 import { url } from "config";
-import { SetUserAction } from "types/redux/actions";
-import { UserGet, UserPatch } from "./typing";
+import { UserGet, UserPatch } from "types/api/api";
+import { SetUserAction } from "types/models/actions";
 
 const handleGetUser: UserGet = async (accessToken) => {
   const response = await fetch(`http://${url}/user`, {
