+++ src/modules/navigation/appbar.tsx
@@ -33,7 +33,7 @@ import { useStyles } from './styles';
 
 // Redux Imports
 import { logoutUser } from 'redux/slices/user';
-import { UserState } from 'types/redux/user';
+import { UserState } from 'types/models/user';
 
 /**
  * A Component that holds the top navigation bar of the application.
