+++ src/modules/realtimegraphs/components.tsx
@@ -15,25 +15,10 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-import { 
-    useStyles
-} from "./styles"
-import { 
-    Switch, 
-    FormControlLabel, 
-    Typography 
-} from '@material-ui/core';
-import {
-    LineChart,
-    Line,
-    XAxis,
-    YAxis,
-    CartesianGrid,
-    Label
-} from "recharts";
-import type {
-    GraphData
-} from "./typing";
+import { useStyles } from "./styles"
+import {  Switch, FormControlLabel, Typography } from '@material-ui/core';
+import { LineChart, Line, XAxis, YAxis, CartesianGrid, Label } from "recharts";
+import { GraphData } from "types/models/ui-types";
 
 export const SensorPaperHeaderHideButton = (props: { onChange: (event: any) => void, checked: boolean }) => {
     const classes = useStyles();
