+++ src/types/models/sensor.d.ts
